This version does not update the top level of the variables that have 'lev' as the vertical coordinate because _some_ of those variables have netCDF MISSING values. The fill_top() routine is required to read the variables in their entirety and replaces the top level with the (valid) data from the level just below the top.
Improvements in bulletproofing - in the scripting and the model_mod.
No more silent failing that I can see.

The vertical interpolation routines work, so model_interpolate is fully
functional (as far as I know).

I removed the estimate_f10_7 from the namelist as per Tomoko's instruction.

We also corrected the situation that when f10_7 is part of the DART state vector,
it should always be updated in the TIEGCM namelist.  

I am still testing with run_filter.csh ... this code should be considered
useful, but should be carefully tested by Tomoko and Chih-Ting.
There are still 'FIXME's for them.

Addresses part of DARTSUP-1



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/tiegcm@7001 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 8a9c9aff5b068689d99ec19e868f9af93f6733ab